requirements.txt
boost==0.1
clang==5.0
h5py==3.1.0
imageio==2.31.2
Jinja2==3.1.2
matplotlib==3.5.3
numpy==1.21.6
open3d==0.17.0
opencv-python==4.8.0.76
pandas==1.3.5
Pillow==9.5.0
plyfile==0.7.4
pyglet==1.5.28
pyrender==0.1.45
pytest==6.2.5
PyYAML==5.4.1
requests-oauthlib==1.0.0
scikit-image==0.18.3
scikit-learn==1.0.2
scipy==1.11.3
sklearn==0.0
tensorboard==2.11.2
tomlkit==0.12.4
tqdm==4.66.1
trimesh==4.1.4
